Separated main service into 2: users and reservations
