+++ reservations/src/main/java/edu/uwb/gymapp/models/Gender.java
@@ -1,0 +1,5 @@
+package edu.uwb.gymapp.models;
+
+public enum Gender {
+    MALE, FEMALE
+}
