+++ reservations/src/main/java/edu/uwb/gymapp/models/Reservation.java
@@ -1,0 +1,91 @@
+package edu.uwb.gymapp.models;
+
+import javax.persistence.*;
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+
+@Entity
+public class Reservation {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Basic(optional = false)
+    @Column(name = "book_time",
+            insertable = false,
+            updatable = false,
+            columnDefinition = "TIMESTAMP DEFAULT CURRENT_TIMESTAMP")
+    private LocalDateTime bookTime;
+
+    @Basic(optional = false)
+    @Column(name = "scheduled_time", updatable = false)
+    private LocalDateTime scheduledTime;
+
+    @Column(name = "cancel_time")
+    private LocalDateTime cancelTime;
+
+    @ManyToOne
+    private Session session;
+
+
+    @Basic(optional = false)
+    @Column(name = "member_email")
+    private String memberEmail;
+
+    public LocalDateTime getScheduledTime() {
+        return scheduledTime;
+    }
+
+    public void setScheduledTime(LocalDateTime scheduledTime) {
+        this.scheduledTime= scheduledTime;
+    }
+
+    public String getMemberEmail() {
+        return memberEmail;
+    }
+
+    public void setMemberEmail(String memberEmail) {
+        this.memberEmail = memberEmail;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public LocalDateTime getBookTime() {
+        return bookTime;
+    }
+
+    public void setBookTime(LocalDateTime bookTime) {
+        this.bookTime = bookTime;
+    }
+
+    public LocalDateTime getCancelTime() {
+        return cancelTime;
+    }
+
+    public void setCancelTime(LocalDateTime cancelTime) {
+        this.cancelTime = cancelTime;
+    }
+
+    public Session getSession() {
+        return session;
+    }
+
+    public void setSession(Session session) {
+        this.session = session;
+    }
+
+//    public Member getMember() {
+//        return member;
+//    }
+//
+//    public void setMember(Member member) {
+//        this.member = member;
+//    }
+}
