+++ reservations/src/main/java/edu/uwb/gymapp/models/SessionRepository.java
@@ -1,0 +1,11 @@
+package edu.uwb.gymapp.models;
+
+import edu.uwb.gymapp.models.Session;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.CrudRepository;
+
+import java.util.List;
+
+public interface SessionRepository extends CrudRepository<Session, Long> {
+
+}
