+++ reservations/src/main/java/edu/uwb/gymapp/resources/MemberDetails.java
@@ -1,0 +1,62 @@
+package edu.uwb.gymapp.resources;
+
+import edu.uwb.gymapp.models.Member;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+
+import java.util.Arrays;
+import java.util.Collection;
+import java.util.List;
+import java.util.stream.Collectors;
+
+public class MemberDetails implements UserDetails {
+
+    private Member member;
+    private List<GrantedAuthority> authorities;
+
+    public MemberDetails(Member member) {
+        this.member = member;
+        String[] strRoles = {"MEMBER"};
+        this.authorities = Arrays.stream(strRoles).map(SimpleGrantedAuthority::new).collect(Collectors.toList());
+    }
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        return authorities;
+    }
+
+    @Override
+    public String getPassword() {
+        return member.getPassword();
+    }
+
+    @Override
+    public String getUsername() {
+        return member.getEmail();
+    }
+
+    @Override
+    public boolean isAccountNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isAccountNonLocked() {
+        return true;
+    }
+
+    @Override
+    public boolean isCredentialsNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isEnabled() {
+        return true;
+    }
+
+    public String getFullName() {
+        return member.getFirstName() + " " + member.getLastName();
+    }
+}
