+++ reservations/src/main/java/edu/uwb/gymapp/workoutsession/SessionController.java
@@ -1,0 +1,20 @@
+package edu.uwb.gymapp.workoutsession;
+
+import edu.uwb.gymapp.models.Session;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.util.List;
+
+@RestController
+public class SessionController {
+
+    @Autowired
+    private SessionService sessionService;
+
+    @RequestMapping("/sessions")
+    public List<Session> getAllSessions() {
+        return sessionService.getAllSessions();
+    }
+}
