--- user-management/src/main/java/edu/uwb/gymapp/UserManagementApplication.java
@@ -1,12 +1,0 @@
-package edu.uwb.gymapp;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class UserManagementApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(UserManagementApplication.class, args);
-	}
-}
