--- user-management/src/main/java/edu/uwb/gymapp/schedule/Reservation.java
@@ -1,71 +1,0 @@
-package edu.uwb.gymapp.schedule;
-
-import edu.uwb.gymapp.usermanagement.Member;
-import edu.uwb.gymapp.workoutsession.Session;
-
-import javax.persistence.*;
-import java.time.LocalDateTime;
-
-@Entity
-public class Reservation {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Basic(optional = false)
-    @Column(name = "book_time",
-            insertable = false,
-            updatable = false,
-            columnDefinition = "TIMESTAMP DEFAULT CURRENT_TIMESTAMP")
-    private LocalDateTime bookTime;
-
-    @Column(name = "cancel_time")
-    private LocalDateTime cancelTime;
-
-    @ManyToOne
-    private Session session;
-
-    @ManyToOne
-    private Member member;
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public LocalDateTime getBookTime() {
-        return bookTime;
-    }
-
-    public void setBookTime(LocalDateTime bookTime) {
-        this.bookTime = bookTime;
-    }
-
-    public LocalDateTime getCancelTime() {
-        return cancelTime;
-    }
-
-    public void setCancelTime(LocalDateTime cancelTime) {
-        this.cancelTime = cancelTime;
-    }
-
-    public Session getSession() {
-        return session;
-    }
-
-    public void setSession(Session session) {
-        this.session = session;
-    }
-
-    public Member getMember() {
-        return member;
-    }
-
-    public void setMember(Member member) {
-        this.member = member;
-    }
-}
