--- user-management/src/main/java/edu/uwb/gymapp/schedule/ReservationRepository.java
@@ -1,12 +1,0 @@
-package edu.uwb.gymapp.schedule;
-
-import org.springframework.data.repository.CrudRepository;
-
-import java.util.List;
-
-public interface ReservationRepository extends CrudRepository<Reservation, Long> {
-
-    public List<Reservation> findByMemberId(Long memberId);
-
-    public List<Reservation> findByMemberEmail(String email);
-}
