--- user-management/src/main/java/edu/uwb/gymapp/usermanagement/Gender.java
@@ -1,5 +1,0 @@
-package edu.uwb.gymapp.usermanagement;
-
-public enum Gender {
-    MALE, FEMALE
-}
