--- user-management/src/main/java/edu/uwb/gymapp/usermanagement/MemberService.java
@@ -1,41 +1,0 @@
-package edu.uwb.gymapp.usermanagement;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
-
-@Service
-public class MemberService {
-
-    @Autowired
-    private MemberRepository memberRepository;
-
-    public List<Member> getAllTopics() {
-        List<Member> members = new ArrayList<>();
-        memberRepository.findAll().forEach(members::add);
-        return members;
-    }
-
-    public Member getMember(Long id) {
-        return memberRepository.findById(id).get();
-    }
-
-    public Member getMember(String username) {
-        return memberRepository.findByUsername(username);
-    }
-
-    public Member addMember(Member member) {
-        return memberRepository.save(member);
-    }
-
-    public void updateMember(Member member) {
-        memberRepository.save(member);
-    }
-
-    public void deleteMember(Long id) {
-        memberRepository.deleteById(id);
-    }
-}
