--- user-management/src/main/java/edu/uwb/gymapp/workoutsession/SessionRepository.java
@@ -1,7 +1,0 @@
-package edu.uwb.gymapp.workoutsession;
-
-import org.springframework.data.repository.CrudRepository;
-
-public interface SessionRepository extends CrudRepository<Session, Long> {
-
-}
