--- user-management/src/main/java/edu/uwb/gymapp/workoutsession/SessionService.java
@@ -1,19 +1,0 @@
-package edu.uwb.gymapp.workoutsession;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@Service
-public class SessionService {
-    @Autowired
-    private SessionRepository sessionRepository;
-
-    public List<Session> getAllSessions() {
-        List<Session> sessions = new ArrayList<>();
-        sessionRepository.findAll().forEach(sessions::add);
-        return sessions;
-    }
-}
