--- user-management/src/main/java/edu/uwb/gymapp/workoutsession/Trainer.java
@@ -1,112 +1,0 @@
-package edu.uwb.gymapp.workoutsession;
-
-import edu.uwb.gymapp.usermanagement.Gender;
-import org.hibernate.annotations.NaturalId;
-
-import javax.persistence.*;
-import java.sql.Timestamp;
-
-@Entity
-public class Trainer {
-    @Id
-    @GeneratedValue(strategy= GenerationType.IDENTITY)
-    private Long id;
-
-    private String username;
-
-    @NaturalId
-    @Column(nullable = false, unique = true)
-    private String email;
-
-    private String password;
-
-    @Column(name = "first_name")
-    private String firstName;
-
-    @Column(name = "last_name")
-    private String lastName;
-
-    @Enumerated(EnumType.STRING)
-    private Gender gender;
-
-    @Basic(optional = false)
-    @Column(name = "create_time",
-            insertable = false,
-            updatable = false,
-            columnDefinition = "TIMESTAMP DEFAULT CURRENT_TIMESTAMP")
-    private java.sql.Timestamp joinDate;
-
-    private String bio;
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-
-    public Gender getGender() {
-        return gender;
-    }
-
-    public void setGender(Gender gender) {
-        this.gender = gender;
-    }
-
-    public Timestamp getJoinDate() {
-        return joinDate;
-    }
-
-    public void setJoinDate(Timestamp joinDate) {
-        this.joinDate = joinDate;
-    }
-
-    public String getBio() {
-        return bio;
-    }
-
-    public void setBio(String bio) {
-        this.bio = bio;
-    }
-}
