+++ users/hosts.bk1
@@ -1,0 +1,13 @@
+#F5 Networks Inc. :File modified by VPN process
+#F5 Networks Inc. :File modified by VPN process
+#F5 Networks Inc. :File modified by VPN process
+#F5 Networks Inc. :File modified by VPN process
+127.0.0.1	localhost
+127.0.1.1	trantor
+::1		localhost
+127.0.1.1	u.washington.edu	trantor	
+
+
+
+
+128.208.179.198 huskyonnet.uw.edu
