+++ users/src/main/java/edu/uwb/gymapp/UserManagementApplication.java
@@ -1,0 +1,17 @@
+package edu.uwb.gymapp;
+
+import edu.uwb.gymapp.models.MemberRepository;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+
+@SpringBootApplication
+@EnableJpaRepositories(basePackageClasses = {MemberRepository.class})
+public class UserManagementApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(UserManagementApplication.class, args);
+	}
+
+
+}
