+++ users/src/main/java/edu/uwb/gymapp/models/MemberRepository.java
@@ -1,0 +1,13 @@
+package edu.uwb.gymapp.models;
+
+import edu.uwb.gymapp.models.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+
+    public Optional<Member> findByUsername(String username);
+
+    public Optional<Member> findByEmail(String email);
+}
