+++ users/src/main/java/edu/uwb/gymapp/resources/MemberDetailsService.java
@@ -1,0 +1,23 @@
+package edu.uwb.gymapp.resources;
+
+import edu.uwb.gymapp.models.Member;
+import edu.uwb.gymapp.models.MemberRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+
+import java.util.Optional;
+
+public class MemberDetailsService implements UserDetailsService {
+
+    @Autowired
+    private MemberRepository memberRepository;
+
+    @Override
+    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
+        Optional<Member> member = memberRepository.findByEmail(email);
+        member.orElseThrow(() -> new UsernameNotFoundException("user not found: " + email));
+        return member.map(MemberDetails::new).get();
+    }
+}
