+++ users/src/main/java/edu/uwb/gymapp/resources/MemberService.java
@@ -1,0 +1,49 @@
+package edu.uwb.gymapp.resources;
+
+import edu.uwb.gymapp.models.Member;
+import edu.uwb.gymapp.models.MemberRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.stereotype.Service;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@Service
+public class MemberService {
+
+    @Autowired
+    private MemberRepository memberRepository;
+
+    public List<Member> getAllMembers() {
+        List<Member> members = new ArrayList<>();
+        memberRepository.findAll().forEach(members::add);
+        return members;
+    }
+
+    public Member getMember(Long id) {
+        return memberRepository.findById(id).get();
+    }
+
+    public Member getMember(String email) {
+        return memberRepository.findByEmail(email).get();
+    }
+
+    public Member addMember(Member member) {
+        BCryptPasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
+        String encodedPassword = passwordEncoder.encode(member.getPassword());
+        member.setPassword(encodedPassword);
+        return memberRepository.save(member);
+    }
+
+    public void updateMember(Member member) {
+        BCryptPasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
+        String encodedPassword = passwordEncoder.encode(member.getPassword());
+        member.setPassword(encodedPassword);
+        memberRepository.save(member);
+    }
+
+    public void deleteMember(Long id) {
+        memberRepository.deleteById(id);
+    }
+}
