+++ users/src/main/java/edu/uwb/gymapp/resources/SecurityConfiguration.java
@@ -1,0 +1,59 @@
+package edu.uwb.gymapp.resources;
+
+import edu.uwb.gymapp.resources.MemberDetailsService;
+import org.springframework.context.annotation.Bean;
+import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
+import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
+import org.springframework.security.config.annotation.web.builders.WebSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+
+@EnableWebSecurity
+public class SecurityConfiguration extends WebSecurityConfigurerAdapter {
+
+//    @Autowired
+//    UserDetailsService userDetailsService;
+
+    @Override
+    protected void configure(AuthenticationManagerBuilder auth) throws Exception {
+        // Set configuration on the auth object
+        auth.userDetailsService(userDetailsService());
+    }
+
+    @Override
+    public void configure(WebSecurity web  /*HttpSecurity http*/) throws Exception {
+        // Set permission
+        web.ignoring().antMatchers("/**");
+
+//        http.authorizeRequests()
+//                .antMatchers("/members?", "members/**").hasAnyRole("MEMBER")
+//                .antMatchers("/").permitAll()
+//                .and().formLogin();
+    }
+
+    @Bean
+    public UserDetailsService userDetailsService() {
+        return new MemberDetailsService();
+    }
+
+    @Bean
+    public BCryptPasswordEncoder passwordEncoder() {
+        return new BCryptPasswordEncoder();
+    }
+
+    @Bean
+    public DaoAuthenticationProvider authenticationProvider() {
+        DaoAuthenticationProvider authProvider = new DaoAuthenticationProvider();
+        authProvider.setUserDetailsService(userDetailsService());
+        authProvider.setPasswordEncoder(passwordEncoder());
+
+        return authProvider;
+    }
+
+//    @Bean
+//    public PasswordEncoder getPasswordEncoder() {
+//        return NoOpPasswordEncoder.getInstance();
+//    }
+}
