+++ users/src/main/resources/application.properties
@@ -1,0 +1,16 @@
+# DATABASE
+spring.jpa.hibernate.ddl-auto=update
+spring.datasource.url=jdbc:mysql://${MYSQL_HOST:jardiamj.ovid.u.washington.edu}:1122/GymMembers
+spring.datasource.username=gymapp
+spring.datasource.password=GymR4t
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+spring.jpa.database-platform=org.hibernate.dialect.MySQL5Dialect
+
+# SERVER PORT
+server.port=18001
+server.error.include-message=always
+
+#ACTUATOR
+management.endpoint.health.show-details=always
+management.endpoints.web.exposure.include=health,beans
+
