+++ users/src/test/java/edu/uwb/gymapp/usermanagement/UserManagementApplicationTests.java
@@ -1,0 +1,13 @@
+package edu.uwb.gymapp.usermanagement;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class UserManagementApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
